requirements.txt
phidata
python-dotenv
packaging
duckduckgo-search
fastapi
uvicorn
groq
slack
